cleaned up syntax
