+++ arena_sport/lib/models.dart
@@ -151,24 +151,3 @@ class HomePageModel extends Model {
     return _leaguesFollowing.map<T>(f);
   }
 }
-
-class FixtureResponse {
-  final int resultCount;
-  final List<Fixture> fixtures;
-
-  FixtureResponse({
-    this.resultCount,
-    this.fixtures,
-  });
-
-  factory FixtureResponse.fromJson(Map<String, dynamic> json) {
-    var data = json['api'];
-    var list = data['fixtures'];
-
-    return FixtureResponse(
-      resultCount: data['results'],
-      fixtures: list.map<Fixture>((json) => Fixture.fromJson(json)).toList(),
-    );
-  }
-}
-
